# specs/003-task-tags/contracts/openapi-tags.yaml
openapi: 3.1.0
info:
  title: Task Tags API Extension
  description: API contract for task tags/categories feature
  version: 1.0.0

paths:
  /api/tasks/tags:
    get:
      summary: List all user tags
      description: |
        Retrieve all unique tags used by the authenticated user across all their tasks.
        Tags are returned sorted alphabetically.
      operationId: listUserTags
      tags:
        - Tags
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of unique tags
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  pattern: '^[a-z0-9-]{1,50}$'
                  example: "work"
              examples:
                typical:
                  summary: Typical tag list
                  value: ["home", "personal", "urgent", "work"]
                empty:
                  summary: User has no tags
                  value: []
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/tasks:
    get:
      summary: List tasks with tag filtering
      description: |
        Retrieve user's tasks with optional filtering by tags.
        When multiple tags are provided, uses AND logic (task must have ALL specified tags).
      operationId: listTasks
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: tags
          in: query
          description: Filter by tags (AND logic - task must have ALL specified tags)
          required: false
          schema:
            type: array
            items:
              type: string
              pattern: '^[a-z0-9-]{1,50}$'
          style: form
          explode: true
          examples:
            single:
              summary: Filter by single tag
              value: ["work"]
            multiple:
              summary: Filter by multiple tags (AND logic)
              value: ["work", "urgent"]
        - name: status
          in: query
          description: Filter by status
          required: false
          schema:
            type: string
            enum: [all, pending, completed, overdue]
        - name: sort_by
          in: query
          description: Sort field
          required: false
          schema:
            type: string
            enum: [due_date, priority, status, created_at]
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: skip
          in: query
          description: Pagination offset
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum results
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTasksResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create task with tags
      description: Create a new task with optional tags
      operationId: createTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            examples:
              withTags:
                summary: Task with tags
                value:
                  title: "Complete project proposal"
                  description: "Write and submit Q1 proposal"
                  due_date: "2025-01-15"
                  priority: "high"
                  tags: ["work", "urgent"]
              withoutTags:
                summary: Task without tags
                value:
                  title: "Buy groceries"
                  due_date: "2025-12-24"
                  tags: []
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/tasks/{task_id}:
    put:
      summary: Update task including tags
      description: Update a task's properties including tags
      operationId: updateTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            examples:
              addTags:
                summary: Add tags to task
                value:
                  tags: ["work", "urgent", "important"]
              removeTags:
                summary: Remove all tags
                value:
                  tags: []
              updateTitle:
                summary: Update title only
                value:
                  title: "Updated task title"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better-Auth

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Complete project proposal"
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: "Write and submit the Q1 project proposal"
        due_date:
          type: string
          format: date
          nullable: true
          example: "2025-01-15"
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        status:
          type: string
          enum: [pending, completed]
          default: pending
        tags:
          type: array
          items:
            type: string
            pattern: '^[a-z0-9-]{1,50}$'
          maxItems: 10
          default: []
          description: Task category labels (max 10, lowercase alphanumeric + hyphens)
          example: ["work", "urgent"]

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, completed]
        tags:
          type: array
          items:
            type: string
            pattern: '^[a-z0-9-]{1,50}$'
          maxItems: 10
          description: Task category labels (max 10, lowercase alphanumeric + hyphens)

    TaskRead:
      type: object
      required:
        - id
        - user_id
        - title
        - priority
        - status
        - tags
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        user_id:
          type: string
          format: uuid
          example: "987e6543-e21b-12d3-a456-426614174000"
        title:
          type: string
          example: "Complete project proposal"
        description:
          type: string
          nullable: true
          example: "Write and submit the Q1 project proposal"
        due_date:
          type: string
          format: date
          nullable: true
          example: "2025-01-15"
        priority:
          type: string
          enum: [low, medium, high]
          example: "high"
        status:
          type: string
          enum: [pending, completed]
          example: "pending"
        tags:
          type: array
          items:
            type: string
            pattern: '^[a-z0-9-]{1,50}$'
          example: ["work", "urgent"]
        created_at:
          type: string
          format: date-time
          example: "2025-12-11T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-12-11T10:30:00Z"

    PaginatedTasksResponse:
      type: object
      required:
        - items
        - total
        - skip
        - limit
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskRead'
        total:
          type: integer
          description: Total count of tasks matching filters
          example: 150
        skip:
          type: integer
          description: Current pagination offset
          example: 0
        limit:
          type: integer
          description: Current pagination limit
          example: 20
        has_more:
          type: boolean
          description: Whether there are more tasks to fetch
          example: true

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Maximum 10 tags allowed per task"

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  loc:
                    type: array
                    items:
                      type: string
                  msg:
                    type: string
                  type:
                    type: string

  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Could not validate credentials"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"

    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          examples:
            tooManyTags:
              summary: Too many tags
              value:
                detail: "Maximum 10 tags allowed per task"
            invalidTagFormat:
              summary: Invalid tag format
              value:
                detail: "Tags can only contain lowercase letters, numbers, and hyphens: 'urgent!!!'"
            emptyTag:
              summary: Empty tag
              value:
                detail: "Tag cannot be empty"